forum_backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0  # 后续部署用，本地开发可不用
slugify
bleach